Apply consistent font to header links & button.  Vertically center header.
The header is using default browser fonts, which was OK but with [New document] now being a button it got different font from the other links.
Experimented with same serif fonts as the editor but felt sans-serif is better; looked at rhythmus's redesign #45 which uses sans and stole his fonts.
[Reminder to self: stop goofing around adding bugs and merge most of his design.]

Also improved vertical centering of header (the float: right half was really off in firefox).
Used simplest height = line-height technique (http://stackoverflow.com/a/8865463/239657) which is not fully robust.