Update DNS zone files reflecting current Cloudflare config. (the files don't do anything, are just exported from Cloudflare UI to keep a record)
They were stale - predate the gh-pages => rhcloud move.
Today's change is an attempt to set a sanely low TTL on CNAME-flattened
apex domains: 5min on mathdown.net, 2min on mathdown.com
(instead of "automatic TTL which is supposed to mean 5min but was
returning *7 days* - #104).
That didn't leave a trace in the zone files and didn't (so war) work...