[test] Bump a the newer of Safari and IE pairs to 'latest'
Safari 10.0 on latest macOS fell off Sauce Labs
support matrix, needed bump to 10.1.
As always I wish to use 'oldest' per each OS,
but that's not supported, need to bump frequently
to ride the back wave...
At least for the browsers I'm running 2 of,
it sounds OK for the one on newer OS to use
latest rather than oldest.